--- .github/pull_request_template.md
@@ -1,3 +1,0 @@
-# What I added (link any issues addressed)
-
-# How to test it
