+++ CribblyBackend/Auth/HasScopeHandler.cs
