+++ CribblyBackend/Auth/HasScopeRequirement.cs
