+++ CribblyBackend/appsettings.json
@@ -1,0 +1,14 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "AllowedHosts": "*",
+  "Auth0": {
+    "Domain": "dev-ib-rt2lw.us.auth0.com",
+    "Audience": "https://cribbly.com/api"
+  }
+}
