--- appsettings.json
@@ -1,14 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "AllowedHosts": "*",
-  "Auth0": {
-    "Domain": "dev-ib-rt2lw.us.auth0.com",
-    "Audience": "https://cribbly.com/api"
-  }
-}
